+++ lib/cldr/backend/formatter.ex
@@ -147,7 +147,8 @@ defmodule Cldr.DateTime.Formatter.Backend do
         defp transliterate(formatted, locale, number_system) do
           with {:ok, number_system} <-
                  Number.System.system_name_from(number_system, locale, unquote(backend)) do
-            Number.Transliterate.transliterate_digits(formatted, :latn, number_system)
+            transliterator = Module.concat(unquote(backend), :"Number.Transliterate")
+            transliterator.transliterate(formatted, :latn, number_system)
           end
         end