+++ lib/cldr/interval/date.ex
@@ -1,5 +1,4 @@
 defmodule Cldr.Date.Interval do
-
   alias Cldr.DateTime.Format
   import Cldr.Calendar, only: [date: 0]
 
@@ -61,15 +60,136 @@ defmodule Cldr.Date.Interval do
     @style_map
   end
 
+  @doc false
+  def to_string(%Date.Range{first: first, last: last}) do
+    locale = Cldr.get_locale
+    backend = locale.backend
+    to_string(first, last, backend, [])
+  end
+
+  if Cldr.Code.ensure_compiled?(CalendarInterval) do
+    @doc false
+    def to_string(%CalendarInterval{} = interval) do
+      locale = Cldr.get_locale
+      backend = locale.backend
+      to_string(interval, backend, [])
+    end
+  end
+
+  @doc false
+  def to_string(unquote(date()) = from, unquote(date()) = to) do
+    locale = Cldr.get_locale
+    backend = locale.backend
+    to_string(from, to, backend, locale: locale)
+  end
+
+  @doc false
   def to_string(%Date.Range{first: first, last: last}, backend) do
     to_string(first, last, backend, [])
   end
 
   if Cldr.Code.ensure_compiled?(CalendarInterval) do
+    @doc false
     def to_string(%CalendarInterval{} = interval, backend) do
       to_string(interval, backend, [])
     end
   end
+
+  @doc false
+  def to_string(unquote(date()) = from, unquote(date()) = to, backend) do
+    to_string(from, to, backend, [])
+  end
+
+  @doc """
+  Returns a `Date.Range` or `CalendarInterval` as
+  a localised string.
+
+  ## Arguments
+
+  * `range` is either a `Date.Range.t` returned from `Date.range/2`
+    or a `CalendarInterval.t`
+
+  * `backend` is any module that includes `use Cldr` and
+    is therefore an `ex_cldr` backend module
+
+  * `options` is a keyword list of options. The default is `[]`.
+
+  ## Options
+
+  * `:format` is one of `:short`, `:medium` or `:long` or a
+    specific format type or a string representing of an interval
+    format. The deault is `:medium`.
+
+  * `:style` supports dfferent formatting styles. The
+    alternatives are `:date`, `:month_and_day`, `:month`
+    and `:year_and_month`. The default is `:date`.
+
+  * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
+    or a `Cldr.LanguageTag` struct.  The default is `Cldr.get_locale/0`
+
+  * `number_system:` a number system into which the formatted date digits should
+    be transliterated
+
+  ## Returns
+
+  * `{:ok, string}` or
+
+  * `{:error, {exception, reason}}`
+
+  ## Notes
+
+  * `CalendarInterval` support requires adding the
+    dependency [calendar_interval](https://hex.pn/packages/calendar_interval)
+    to the `deps` configuration in `mix.exs`.
+
+  * For more information on interval format string
+    see the [module documentation]().
+
+  * The available predefined formats that can be applied are the
+    keys of the map returned by `Cldr.DateTime.Format.interval_formats("en", :gregorian)
+    where `"en"` can be replaced by any configuration locale name and `:gregorian`
+    is the underlying `CLDR` calendar type.
+
+  * In the case where `from` and `to` are equal, a single
+    date is formatted instead of an interval
+
+  ## Examples
+
+      iex> Cldr.Date.Interval.to_string Date.range(~D[2020-01-01], ~D[2020-12-31]), MyApp.Cldr
+      {:ok, "Jan 1 – Dec 31, 2020"}
+
+      iex> Cldr.Date.Interval.to_string Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr
+      {:ok, "Jan 1 – 12, 2020"}
+
+      iex> Cldr.Date.Interval.to_string Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr,
+      ...> format: :long
+      {:ok, "Wed, Jan 1 – Sun, Jan 12, 2020"}
+
+      iex> Cldr.Date.Interval.to_string Date.range(~D[2020-01-01], ~D[2020-12-01]), MyApp.Cldr,
+      ...> format: :long, style: :year_and_month
+      {:ok, "January – December 2020"}
+
+      iex> use CalendarInterval
+      iex> Cldr.Date.Interval.to_string ~I"2020-01/12", MyApp.Cldr
+      {:ok, "Jan 1 – Dec 31, 2020"}
+
+      iex> Cldr.Date.Interval.to_string Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr,
+      ...> format: :short
+      {:ok, "1/1/2020 – 1/12/2020"}
+
+      iex> Cldr.Date.Interval.to_string Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr,
+      ...> format: :long, locale: "fr"
+      {:ok, "mer. 1 – dim. 12 janv. 2020"}
+
+  """
+  if Cldr.Code.ensure_compiled?(CalendarInterval) do
+    @type range :: Date.Range.t() | CalendarInterval.t()
+  else
+    @type Date.Range.t()
+  end
+
+  @spec to_string(range, Cldr.backend(), Keyword.t()) ::
+      {:ok, String.t()}, {:error, {module(), String.t()}}
 
   def to_string(%Date.Range{first: first, last: last}, backend, options) do
     to_string(first, last, backend, options)
@@ -89,6 +209,92 @@ defmodule Cldr.Date.Interval do
     end
   end
 
+  @doc """
+  Returns a string representing the formatted
+  interval formed by two date.
+
+  ## Arguments
+
+  * `from` and `to` are any maps that conform to the
+    `Calendar.date` type which means a map that includes
+    at least the keys `:year`, `:month` and `:day`
+
+  * `backend` is any module that includes `use Cldr` and
+    is therefore an `ex_cldr` backend module
+
+  * `options` is a keyword list of options. The default is `[]`.
+
+  ## Options
+
+  * `:format` is one of `:short`, `:medium` or `:long` or a
+    specific format type or a string representing of an interval
+    format. The deault is `:medium`.
+
+  * `:style` supports dfferent formatting styles. The
+    alternatives are `:date`, `:month_and_day`, `:month`
+    and `:year_and_month`. The default is `:date`.
+
+  * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
+    or a `Cldr.LanguageTag` struct.  The default is `Cldr.get_locale/0`
+
+  * `number_system:` a number system into which the formatted date digits should
+    be transliterated
+
+  ## Returns
+
+  * `{:ok, string}` or
+
+  * `{:error, {exception, reason}}`
+
+  ## Notes
+
+  * `CalendarInterval` support requires adding the
+    dependency [calendar_interval](https://hex.pn/packages/calendar_interval)
+    to the `deps` configuration in `mix.exs`.
+
+  * For more information on interval format string
+    see the [module documentation]().
+
+  * The available predefined formats that can be applied are the
+    keys of the map returned by `Cldr.DateTime.Format.interval_formats("en", :gregorian)
+    where `"en"` can be replaced by any configuration locale name and `:gregorian`
+    is the underlying `CLDR` calendar type.
+
+  * In the case where `from` and `to` are equal, a single
+    date is formatted instead of an interval
+
+  ## Examples
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], ~D[2020-12-31], MyApp.Cldr
+      {:ok, "Jan 1 – Dec 31, 2020"}
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr
+      {:ok, "Jan 1 – 12, 2020"}
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :long
+      {:ok, "Wed, Jan 1 – Sun, Jan 12, 2020"}
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], ~D[2020-12-01], MyApp.Cldr,
+      ...> format: :long, style: :year_and_month
+      {:ok, "January – December 2020"}
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :short
+      {:ok, "1/1/2020 – 1/12/2020"}
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :long, locale: "fr"
+      {:ok, "mer. 1 – dim. 12 janv. 2020"}
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :long, locale: "th", number_system: :thai
+      {:ok, "พ. ๑ ม.ค. – อา. ๑๒ ม.ค. ๒๐๒๐"}
+
+  """
+  @spec to_string(Calendar.date(), Calendar.date(), Cldr.backend(), Keyword.t()) ::
+      {:ok, String.t()}, {:error, {module(), String.t()}}
+
   def to_string(from, to, backend, options \\ [])
 
   def to_string(unquote(date()) = from, unquote(date()) = to, backend, options)
@@ -104,9 +310,22 @@ defmodule Cldr.Date.Interval do
     formatter = Module.concat(backend, DateTime.Formatter)
     format = Keyword.get(options, :format, @default_format)
 
+    number_system =
+      Keyword.get(
+        options,
+        :number_system,
+        Cldr.Number.System.number_system_from_locale(locale, backend)
+      )
+
+    options =
+      options
+      |> Keyword.put(:locale, locale)
+      |> Keyword.put(:nunber_system, number_system)
+
     with {:ok, _} <- from_less_than_or_equal_to(from, to),
          {:ok, backend} <- Cldr.validate_backend(backend),
          {:ok, locale} <- Cldr.validate_locale(locale, backend),
+         {:ok, _} <- Cldr.Number.validate_number_system(locale, number_system, backend),
          {:ok, calendar} <- Cldr.Calendar.validate_calendar(from.calendar),
          {:ok, formats} <- Format.interval_formats(locale, calendar.cldr_calendar_type, backend),
          {:ok, [left, right]} <- resolve_format(from, to, formats, options),
@@ -123,9 +342,103 @@ defmodule Cldr.Date.Interval do
     end
   end
 
-  def to_string!(%Date.Range{} = range, backend) do
-    to_string!(range, backend, [])
+  @doc false
+  def to_string!(%Date.Range{first: first, last: last}, backend) do
+    to_string!(first, last, backend, [])
   end
+
+  if Cldr.Code.ensure_compiled?(CalendarInterval) do
+    @doc false
+    def to_string!(%CalendarInterval{} = interval, backend) do
+      to_string!(interval, backend, [])
+    end
+  end
+
+  @doc """
+  Returns a `Date.Range` or `CalendarInterval` as
+  a localised string.
+
+  ## Arguments
+
+  * `range` is either a `Date.Range.t` returned from `Date.range/2`
+    or a `CalendarInterval.t`
+
+  * `backend` is any module that includes `use Cldr` and
+    is therefore an `ex_cldr` backend module
+
+  * `options` is a keyword list of options. The default is `[]`.
+
+  ## Options
+
+  * `:format` is one of `:short`, `:medium` or `:long` or a
+    specific format type or a string representing of an interval
+    format. The deault is `:medium`.
+
+  * `:style` supports dfferent formatting styles. The
+    alternatives are `:date`, `:month_and_day`, `:month`
+    and `:year_and_month`. The default is `:date`.
+
+  * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
+    or a `Cldr.LanguageTag` struct.  The default is `Cldr.get_locale/0`
+
+  * `number_system:` a number system into which the formatted date digits should
+    be transliterated
+
+  ## Returns
+
+  * `string` or
+
+  * raises an exception
+
+  ## Notes
+
+  * `CalendarInterval` support requires adding the
+    dependency [calendar_interval](https://hex.pn/packages/calendar_interval)
+    to the `deps` configuration in `mix.exs`.
+
+  * For more information on interval format string
+    see the [module documentation]().
+
+  * The available predefined formats that can be applied are the
+    keys of the map returned by `Cldr.DateTime.Format.interval_formats("en", :gregorian)
+    where `"en"` can be replaced by any configuration locale name and `:gregorian`
+    is the underlying `CLDR` calendar type.
+
+  * In the case where `from` and `to` are equal, a single
+    date is formatted instead of an interval
+
+  ## Examples
+
+      iex> Cldr.Date.Interval.to_string! Date.range(~D[2020-01-01], ~D[2020-12-31]), MyApp.Cldr
+      "Jan 1 – Dec 31, 2020"
+
+      iex> Cldr.Date.Interval.to_string! Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr
+      "Jan 1 – 12, 2020"
+
+      iex> Cldr.Date.Interval.to_string! Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr,
+      ...> format: :long
+      "Wed, Jan 1 – Sun, Jan 12, 2020"
+
+      iex> Cldr.Date.Interval.to_string! Date.range(~D[2020-01-01], ~D[2020-12-01]), MyApp.Cldr,
+      ...> format: :long, style: :year_and_month
+      "January – December 2020"
+
+      iex> use CalendarInterval
+      iex> Cldr.Date.Interval.to_string! ~I"2020-01/12", MyApp.Cldr
+      "Jan 1 – Dec 31, 2020"
+
+      iex> Cldr.Date.Interval.to_string! Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr,
+      ...> format: :short
+      "1/1/2020 – 1/12/2020"
+
+      iex> Cldr.Date.Interval.to_string! Date.range(~D[2020-01-01], ~D[2020-01-12]), MyApp.Cldr,
+      ...> format: :long, locale: "fr"
+      "mer. 1 – dim. 12 janv. 2020"
+
+  """
+
+  @spec to_string!(range, Cldr.backend(), Keyword.t()) ::
+    String.t() | no_return()
 
   def to_string!(%Date.Range{} = range, backend, options) do
     case to_string(range, backend, options) do
@@ -133,6 +446,101 @@ defmodule Cldr.Date.Interval do
       {:error, {exception, reason}} -> raise exception, reason
     end
   end
+
+  if Cldr.Code.ensure_compiled?(CalendarInterval) do
+    def to_string!(%CalendarInterval{first: from, last: to, precision: precision}, backend, options)
+        when precision in [:year, :month, :day] do
+      to_string!(from, to, backend, options)
+    end
+  end
+
+
+  @doc """
+  Returns a string representing the formatted
+  interval formed by two date.
+
+  ## Arguments
+
+  * `from` and `to` are any maps that conform to the
+    `Calendar.date` type which means a map that includes
+    at least the keys `:year`, `:month` and `:day`. The
+    date `from` must precede or equal the date `to`.
+
+  * `backend` is any module that includes `use Cldr` and
+    is therefore an `ex_cldr` backend module
+
+  * `options` is a keyword list of options
+
+  ## Options
+
+  * `:format` is one of `:short`, `:medium` or `:long` or a
+    specific format type or a string representing of an interval
+    format. The deault is `:medium`.
+
+  * `:style` supports dfferent formatting styles. The
+    alternatives are `:date`, `:month_and_day`, `:month`
+    and `:year_and_month`. The default is `:date`.
+
+  * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
+    or a `Cldr.LanguageTag` struct.  The default is `Cldr.get_locale/0`
+
+  * `number_system:` a number system into which the formatted date digits should
+    be transliterated
+
+  ## Returns
+
+  * `{:ok, string}` or
+
+  * `{:error, {exception, reason}}`
+
+  ## Notes
+
+  * `CalendarInterval` support requires adding the
+    dependency [calendar_interval](https://hex.pn/packages/calendar_interval)
+    to the `deps` configuration in `mix.exs`.
+
+  * For more information on interval format string
+    see the [module documentation]().
+
+  * The available predefined formats that can be applied are the
+    keys of the map returned by `Cldr.DateTime.Format.interval_formats("en", :gregorian)
+    where `"en"` can be replaced by any configuration locale name and `:gregorian`
+    is the underlying `CLDR` calendar type.
+
+  * In the case where `from` and `to` are equal, a single
+    date is formatted instead of an interval
+
+  ## Examples
+
+      iex> Cldr.Date.Interval.to_string! ~D[2020-01-01], ~D[2020-12-31], MyApp.Cldr
+      "Jan 1 – Dec 31, 2020"
+
+      iex> Cldr.Date.Interval.to_string! ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr
+      "Jan 1 – 12, 2020"
+
+      iex> Cldr.Date.Interval.to_string! ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :long
+      "Wed, Jan 1 – Sun, Jan 12, 2020"
+
+      iex> Cldr.Date.Interval.to_string! ~D[2020-01-01], ~D[2020-12-01], MyApp.Cldr,
+      ...> format: :long, style: :year_and_month
+      "January – December 2020"
+
+      iex> Cldr.Date.Interval.to_string! ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :short
+      "1/1/2020 – 1/12/2020"
+
+      iex> Cldr.Date.Interval.to_string! ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :long, locale: "fr"
+      "mer. 1 – dim. 12 janv. 2020"
+
+      iex> Cldr.Date.Interval.to_string! ~D[2020-01-01], ~D[2020-01-12], MyApp.Cldr,
+      ...> format: :long, locale: "th", number_system: :thai
+      "พ. ๑ ม.ค. – อา. ๑๒ ม.ค. ๒๐๒๐"
+
+  """
+  @spec to_string!(Calendar.date(), Calendar.date(), Cldr.backend(), Keyword.t()) ::
+    String.t() | no_return()
 
   def to_string!(from, to, backend, options \\ []) do
     case to_string(from, to, backend, options) do
@@ -243,7 +651,7 @@ defmodule Cldr.Date.Interval do
     {
       Cldr.DateTime.DateTimeOrderError,
       "Start date/time must be earlier or equal to end date/time. " <>
-      "Found #{inspect from}, #{inspect to}."
+        "Found #{inspect(from)}, #{inspect(to)}."
     }
   end
 
@@ -252,7 +660,7 @@ defmodule Cldr.Date.Interval do
     {
       Cldr.DateTime.IncompatibleTimeZonerError,
       "Start and end dates must be in the same time zone. " <>
-      "Found #{inspect from}, #{inspect to}."
+        "Found #{inspect(from)}, #{inspect(to)}."
     }
   end