+++ mix/my_app_backend.ex
@@ -1,6 +1,7 @@
 defmodule MyApp.Cldr do
   use Cldr,
-    locales: ["en", "fr", "af", "ja", "de", "pl"],
+    locales: ["en", "fr", "af", "ja", "de", "pl", "th"],
     providers: [Cldr.Number, Cldr.Calendar, Cldr.DateTime],
-    precompile_number_formats: ["#,##0"]
+    precompile_number_formats: ["#,##0"],
+    precompile_transliterations: [{:latn, :thai}]
 end