+++ test/doc_test.exs
@@ -8,8 +8,17 @@ defmodule Cldr.DateTime.Test do
   doctest Cldr.Date
   doctest Cldr.Time
 
+  doctest Cldr.Interval
+  doctest Cldr.DateTime.Interval
+  doctest Cldr.Date.Interval
+  doctest Cldr.Time.Interval
+
   doctest MyApp.Cldr.Date
   doctest MyApp.Cldr.Time
   doctest MyApp.Cldr.DateTime
   doctest MyApp.Cldr.DateTime.Relative
+
+  doctest MyApp.Cldr.DateTime.Interval
+  doctest MyApp.Cldr.Date.Interval
+  doctest MyApp.Cldr.Time.Interval
 end